refactor [James]: refactor cart routes
